google auhentication created

+++ controllers/api/v1/posts_api.js
@@ -18,11 +18,18 @@ module.exports.index = async function(req, res) {
 module.exports.destroy = async function(req, res) {
     try {
         let post = await Post.findById(req.params.id);
-        post.remove();
-        await Comment.deleteMany({ post: req.params.id });
-        return res.json(200, {
-            message: "Posts and associated comments are deleted"
-        });
+        if (post.user == req.user.id) {
+
+            post.remove();
+            await Comment.deleteMany({ post: req.params.id });
+            return res.json(200, {
+                message: "Posts and associated comments are deleted"
+            });
+        } else {
+            return res.json(404, {
+                message: "You can not delete this post"
+            })
+        }
     } catch (err) {
         return res.json(500, {
             message: "Internal Server Error"
